requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
